+++ CHANGES.rst
@@ -2,7 +2,7 @@ Changelog of threedi-modelchecker
 =================================
 
 
-2.7.1 (unreleased)
+2.7.1 (2024-04-22)
 ------------------
 - Add info check 1406 to inform the user if a raster is not compressed.
 - Add check 799 to warn if raster friction pixels are < 1 while Chezy friction is selected